fix(build): Try fixing the -Wimplicit-int-conversion warnings again
This was tried in 9ddafd03501c8688ccd579585585bf9725c2ebf3 (with a mistake
of mine) and then reverted in 7afada125b3c263af11cda5cbf121ab2dfc5cf60.

Hopefully this new attempt is right; I just read the context again
and it seems fine, now.

+++ src/ScummFont/scummfont.cpp
@@ -612,7 +612,7 @@ static void loadFont(const char *path)
 				p += bpp;
 				if (mask == 0)
 				{
-					mask = ((1 << bpp) - 1) << (8 - bpp);
+					mask = (byte)(((1 << bpp) - 1) << (8 - bpp));
 					b = (byte)file.get();
 					p = 0;
 				}
