+++ src/ScummRp/block.cpp
@@ -2066,11 +2066,11 @@ void OldRoom::_getOIInfo(uint16 bmLastOffset, std::vector<uint16> &oiOffset, con
 
 		_file->seekg(ocOffset[i] + _oOCWidth(), std::ios::beg);
 		_file->getByte(b);
-		width = b << 3;
+		width = (uint16)(b << 3);
 
 		_file->seekg(ocOffset[i] + _oOCHeight(), std::ios::beg);
 		_file->getByte(b);
-		height = b & 0xF8;
+		height = (uint16)(b & 0xF8);
 
 		if (oiOffset[i] >= firstOCOffset || oiOffset[i] <= bmLastOffset) // (1)
 		{
